requirements.txt
crewai>=0.1.0
pyyaml>=6.0.1
pydantic>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
tabulate>=0.9.0  # For better DataFrame display